fix: :bug: fix issues from code review
