+++ src/types/global.d.ts
@@ -1,18 +1,19 @@
 type NodeEnvironment = 'development' | 'local' | 'production';
 
 interface EnvVars {
-	database: D1Database;
-	SESSION_TOKENS: KVNamespace;
-	ACTIVATION_TOKENS: KVNamespace;
 	BASE_URL: string;
 	SENDER_EMAIL: string;
 	SENDGRID_API_KEY: string;
-	ASSETS: Fetcher;
 	NODE_ENV: NodeEnvironment;
 }
 
 interface EnvironmentBindings {
-	Bindings: EnvVars;
+	Bindings: EnvVars & {
+		database: D1Database,
+		SESSION_TOKENS: KVNamespace,
+		ACTIVATION_TOKENS: KVNamespace,
+		ASSETS: Fetcher
+	};
 }
 
 type ISODate = string;
